hutddf

+++ resources/views/admin/roles/index.blade.php
@@ -1,7 +1,7 @@
 @extends('layouts.app')
 
 @section('content')
-    <div class="row">
+   <div class="row">
         <div class="col-lg-12 margin-tb">
             <div class="pull-left">
                 <h2>Role Management</h2>
@@ -15,21 +15,12 @@
     </div>
 
     @if ($message = Session::get('success'))
-
-        <div class="alert alert-success border-0 bg-grd-success alert-dismissible fade show">
-            <div class="d-flex align-items-center">
-                <div class="font-35 text-white"><span class="material-icons-outlined fs-2">check_circle</span>
-                </div>
-                <div class="ms-3">
-                    <h6 class="mb-0 text-white">Success Alerts</h6>
-                    <div class="text-white">A simple success alert—check it out!</div>
-                </div>
-            </div>
-            <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
+        <div class="alert alert-success">
+            <p>{{ $message }}</p>
         </div>
     @endif
 
-    <table class="manage-table resumes responsive-table">
+    <table class="table table-bordered">
         <tr>
             <th>No</th>
             <th>Name</th>
@@ -40,19 +31,20 @@
                 <td>{{ ++$i }}</td>
                 <td>{{ $role->name }}</td>
                 <td>
-                    <a class="btn btn-info" href="{{ route('roles.show', $role->id) }}"><i class="fa fa-eye"></i> Show</a>
+                    <a class="btn btn-info" href="{{ route('roles.show', $role->id) }}">Show</a>
                     @can('role-edit')
-                        <a class="btn btn-primary" href="{{ route('roles.edit', $role->id) }}"><i class="fa fa-pencil"></i>Edit</a>
+                        <a class="btn btn-primary" href="{{ route('roles.edit', $role->id) }}">Edit</a>
                     @endcan
-                    {{-- @can('role-delete')
+                    @can('role-delete')
                         {!! Form::open(['method' => 'DELETE', 'route' => ['roles.destroy', $role->id], 'style' => 'display:inline']) !!}
                         {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
                         {!! Form::close() !!}
-                    @endcan --}}
+                    @endcan
                 </td>
             </tr>
         @endforeach
     </table>
 
     {!! $roles->render() !!}
+
 @endsection
